Refactored main. Put init function in for loading scene for each state
