+++ src/components/render_component.cpp
@@ -6,7 +6,7 @@
 #include <glm/gtc/matrix_transform.hpp>
 #include "camera_component.h"
 #include "../systems/camera_system.h"
-#include "../renderer.h"
+#include "../systems/renderer.h"
 
 #define CHECK_GL_ERROR CheckGL(__LINE__, __FILE__)
 
@@ -43,7 +43,7 @@ void render_component::render()
 		// "Generate" the transform matrix.
 		vec3 transvec = vec3(_parent->get_trans().x, _parent->get_trans().y, _parent->get_trans().z);
 
-		std::cout << "rendering the: " + _parent->get_name() << " at (" << transvec.x << ", " << transvec.y << ", " << transvec.z << ")" << std::endl;
+		//std::cout << "rendering the: " + _parent->get_name() << " at (" << transvec.x << ", " << transvec.y << ", " << transvec.z << ")" << std::endl;
 
 		mat4 trans = glm::translate(mat4(1.0f), transvec);
