+++ src/engine_state.h
@@ -1,5 +1,11 @@
 #pragma once
 #include <iostream>
+#include "entity_manager.h"
+#include "systems/renderer.h"
+#include "systems/physics_system.h"
+#include "systems/input_handler.h"
+#include "systems/clickable_system.h"
+#include "systems/camera_system.h"
 
 enum state_type
 {
@@ -19,6 +25,8 @@ public:
     state_type type;
 
     virtual ~engine_state() = default;
+
+	virtual void initialise() = 0;
 
     virtual void on_enter() = 0;
