+++ src/entity.h
@@ -5,9 +5,10 @@
 #include <memory>
 #include "components/component.h"
 #include <iostream>
-#include "engine_state.h"
 #include <glm/glm.hpp>
 #include <glm/gtx/quaternion.hpp>
+
+enum state_type;
 
 struct transform_data
 {
@@ -32,7 +33,7 @@ private:
 
 public:
 	// Used to keep track of which state this entity belongs to
-	state_type state = state_type::NONE;
+	state_type state;// = state_type::NONE;
 	inline entity(const std::string& name, transform_data trans)
 		: _name(name), _trans(trans)
 	{
