+++ src/game_state.cpp
@@ -1,5 +1,46 @@
 #include "game_state.h"
 #include <iostream>
+
+
+void game_state::initialise()
+{
+
+	// can add a transformation in using struct now
+	auto light = entity_manager::get()->create_entity("Light", this->type);
+	light->add_component("light", renderer::get()->build_light(light));
+
+	// Adding plane
+	auto p = entity_manager::get()->create_entity("Plane", this->type);
+	p->add_component("physics", physics_system::get()->build_component(p));
+	p->add_component("render", renderer::get()->build_component(p, "Red", "", "plane", "Gouraud", simple));
+
+	transform_data cubeTrans;
+	cubeTrans.scale = glm::vec3(10.0f, 10.0f, 10.0f);
+	cubeTrans.x = -20.0f;
+	//cubeTrans.y = 1.0f;
+	cubeTrans.z = -20.0f;
+	// Adding cube
+	auto c = entity_manager::get()->create_entity("Cube", this->type, cubeTrans);
+	c->add_component("physics", physics_system::get()->build_component(c));
+	c->add_component("render", renderer::get()->build_component(c, "Green", "res/textures/check.jpg", "cube", "Gouraud", phong));
+
+	//c->add_component("collider", physics_system::get()->build_collider_component(c, cubeTrans.scale));
+
+	transform_data batTrans;
+	batTrans.y = 5.0f;
+	batTrans.scale = glm::vec3(0.5f, 0.5f, 0.5f);
+	batTrans.rotation = glm::angleAxis(90.0f, glm::vec3(1.0f, 0.0f, 0.0f)) * glm::angleAxis(radians(180.0f), glm::vec3(0.0f, 1.0f, 0.0f));
+
+	// Bat entity
+	auto e = entity_manager::get()->create_entity("Bat", this->type, batTrans);
+	e->add_component("physics", physics_system::get()->build_component(e));
+	e->add_component("input", input_handler::get()->build_component(e));
+	e->add_component("render", renderer::get()->build_component(e, "Blue", "res/textures/bat.jpg", "res/models/bat.obj", "Gouraud", phong));
+	e->add_component("camera", camera_system::get()->build_component(e, camera_type::CHASE));
+	//e->add_component("collider", physics_system::get()->build_collider_component(e, batTrans.scale * 5.0f));
+
+}
+
 void game_state::on_enter()
 {
     // Turn on entities, physics, renderer
@@ -8,6 +49,7 @@ void game_state::on_enter()
     engine::get()->get_subsystem("physics_system")->set_active(true);
     engine::get()->get_subsystem("renderer")->set_visible(true);
     engine::get()->get_subsystem("clickable_system")->set_active(false);
+	;
 
     glfwSetInputMode(glfw::window, GLFW_CURSOR, GLFW_CURSOR_HIDDEN);
