+++ src/main.cpp
@@ -5,17 +5,12 @@
 #define GLM_ENABLE_EXPERIMENTAL
 #include <iostream>
 #include "engine.h"
-#include "entity_manager.h"
-#include "systems/physics_system.h"
-#include "renderer.h"
 #include "state_machine.h"
 #include "engine_state_machine.h"
 #include "game_state.h"
 #include "menu_state.h"
 #include "pause_state.h"
-#include "systems/input_handler.h"
-#include "systems/clickable_system.h"
-#include "systems/camera_system.h"
+
 //#include <FreeImage-src\FreeImage.h>
 // Include GLEW
 #include <GL/glew.h>
@@ -48,74 +43,6 @@ int main()
     engine_state_machine::get()->add_state("game_state", std::make_shared<game_state>(), state_type::GAME);
     engine_state_machine::get()->add_state("pause_state", std::make_shared<pause_state>(), state_type::PAUSE);
     engine_state_machine::get()->change_state("game_state");
-
-	// can add a transformation in using struct now
-	auto light = entity_manager::get()->create_entity("Light", state_type::GAME);
-	light->add_component("light", renderer::get()->build_light(light));
-
-    // Adding plane
-	auto p = entity_manager::get()->create_entity("Plane", state_type::GAME);
-    p->add_component("physics", physics_system::get()->build_component(p));
-    p->add_component("render", renderer::get()->build_component(p, "Red", "", "plane", "Gouraud", simple));
-
-	transform_data cubeTrans;
-	cubeTrans.scale = glm::vec3(10.0f, 10.0f, 10.0f);
-	cubeTrans.x = -20.0f;
-	//cubeTrans.y = 1.0f;
-	cubeTrans.z = -20.0f;
-    // Adding cube
-    auto c = entity_manager::get()->create_entity("Cube", state_type::GAME, cubeTrans);
-    c->add_component("physics", physics_system::get()->build_component(c));
-    c->add_component("render", renderer::get()->build_component(c, "Green", "res/textures/check.jpg", "cube", "Gouraud", phong));
-
-	//c->add_component("collider", physics_system::get()->build_collider_component(c, cubeTrans.scale));
-
-	transform_data batTrans;
-	batTrans.y = 5.0f;
-	batTrans.scale = glm::vec3(0.5f, 0.5f, 0.5f);
-	batTrans.rotation = glm::angleAxis(90.0f, glm::vec3(1.0f, 0.0f, 0.0f)) * glm::angleAxis(radians(180.0f), glm::vec3(0.0f, 1.0f, 0.0f));
-
-    // Bat entity
-    auto e = entity_manager::get()->create_entity("Bat", state_type::GAME, batTrans);
-    e->add_component("physics", physics_system::get()->build_component(e));
-    e->add_component("input", input_handler::get()->build_component(e));
-    e->add_component("render", renderer::get()->build_component(e, "Blue","res/textures/bat.jpg", "res/models/bat.obj", "Gouraud", phong));
-    e->add_component("camera", camera_system::get()->build_component(e, camera_type::CHASE));
-	//e->add_component("collider", physics_system::get()->build_collider_component(e, batTrans.scale * 5.0f));
-
-    int x_size = 0;
-    int y_size = 0;
-
-    glfwGetWindowSize(glfw::window, &x_size, &y_size);
-
-    int x_center = x_size / 2;
-    int y_center = y_size / 2; 
-
-    int x_button_size = 100;
-    int y_button_size = 50;
-
-	// button trans
-	transform_data buttonTrans;
-	buttonTrans.scale = glm::vec3(x_button_size, y_button_size, 1.0f);
-	  
-    // Y offset between buttons
-    int button_offset = 100;
-    // Menu buttons
-    auto button_play = entity_manager::get()->create_entity("buttonPlay", state_type::MENU, buttonTrans);
-    button_play->add_component("clickable", clickable_system::get()->build_component(button_play, glm::vec2(x_center - x_button_size, x_center + x_button_size), glm::vec2(y_center - y_button_size, y_center + y_button_size)));
-    button_play->add_component("render", renderer::get()->build_component(button_play, "Blue", "", "rectangle", "Gouraud", simple));
-	button_play->add_component("camera", camera_system::get()->build_component(button_play, camera_type::ORTHO));
-
-    auto button_exit = entity_manager::get()->create_entity("buttonExit", state_type::MENU);
-    button_exit->add_component("clickable", clickable_system::get()->build_component(button_exit, glm::vec2(x_center - x_button_size, x_center + x_button_size), glm::vec2(button_offset + y_center - y_button_size, button_offset + y_center + y_button_size)));
-
-    // Pause buttons
-    auto button_continue = entity_manager::get()->create_entity("buttonContinue", state_type::PAUSE);
-    button_continue->add_component("clickable", clickable_system::get()->build_component(button_continue, glm::vec2(x_center - x_button_size, x_center + x_button_size), glm::vec2(y_center - y_button_size, y_center + y_button_size)));
-
-    auto button_menu = entity_manager::get()->create_entity("buttonMenu", state_type::PAUSE);
-    button_menu->add_component("clickable", clickable_system::get()->build_component(button_menu, glm::vec2(x_center - x_button_size, x_center + x_button_size), glm::vec2(button_offset + y_center - y_button_size, button_offset + y_center + y_button_size)));
-
     eng->run();
 
     return 0;
