+++ src/menu_state.h
@@ -5,6 +5,8 @@
 class menu_state : public engine_state
 {
 public:
+	void initialise();
+
     void on_enter();
 
     void on_update(float delta_time);
