+++ src/start_state.h
@@ -5,6 +5,8 @@
 class start_state : public engine_state
 {
 public:
+	void initialise();
+
     void on_enter();
 
     void on_update(float delta_time);
