+++ src/systems/renderer.cpp
@@ -2,7 +2,7 @@
 #include "renderer.h"
 #include <sstream>
 #include <fstream>
-#include "glfw.h"
+#include "../glfw.h"
 #include <glm/glm.hpp>
 #include <glm/gtc/type_ptr.hpp>
 #include <glm/gtc/matrix_transform.hpp>
